+++ src/iec-types.ts
@@ -22,635 +22,160 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
-import iconv from 'iconv-lite'
-
 import type {
   IecType,
   StructChildren
 } from './types/types'
 
-
-/**
- * Base abstract type
- */
-abstract class TypeBase implements Partial<IecType> {
-  type = ''
-  byteLength = 0
-}
-
-
-
-
+import * as handler from './iec-type-handler'
 
 /**
  * IEC 61131-3 type: STRUCT
+ * Provide struct children as object
  */
-export class STRUCT extends TypeBase implements IecType {
-  type = 'STRUCT'
-  /**
-   * STRUCT children data types
-   */
-  children: StructChildren
-
-  constructor(children: StructChildren) {
-    super()
-    
-    this.children = children
-
-    //Calculating struct size
-    for (const key in this.children) {
-      this.byteLength += this.children[key].byteLength
-    }
-  }
-
-
-  convertToBuffer(data: StructChildren): Buffer {
-    if (!data)
-      return Buffer.alloc(0)
-
-    const buffer = Buffer.alloc(this.byteLength)
-    let pos = 0
-
-    for (const key in this.children) {
-      const converted = this.children[key].convertToBuffer(data[key])
-
-      converted.copy(buffer, pos)
-      pos += converted.byteLength
-    }
-
-    return buffer
-  }
-
-  convertFromBuffer(data: Buffer): Record<string, unknown> {
-    const obj = {} as Record<string, unknown>
-    let pos = 0
-
-    for (const key in this.children) {
-      obj[key] = this.children[key].convertFromBuffer(data.slice(pos, pos + this.children[key].byteLength))
-      pos += this.children[key].byteLength
-    }
-
-    return obj
-  }
-
-  getDefault(): Record<string, unknown> {
-    const obj = {} as Record<string, unknown>
-
-    for (const key in this.children) {
-      obj[key] = this.children[key].getDefault()
-    }
-
-    return obj
-  }
-}
-
-
-
+export const STRUCT = (children: StructChildren): handler.STRUCT => new handler.STRUCT(children)
 
 /**
  * IEC 61131-3 type: ARRAY
  * Handles 1..3 dimensional arrays
+ *
+ * @param dataType Data type of the array (example: iec.INT)
+ * @param dimensions If 1-dimensional array: Array dimension (size) as number. If multi-dimensional array, array dimensions as array (like [1, 10, 5])
  */
-export class ARRAY extends TypeBase implements IecType {
-  type = 'ARRAY'
-  dataType: IecType
-  dimensions: number[] = []
-  totalSize = 0
-
-  /**
-   * Constructor for array
-   * @param dataType Data type of the array (example: iec.INT)
-   * @param dimensions If 1-dimensional array: Array dimension (size) as number. If multi-dimensional array, array dimensions as array (like [1, 10, 5])
-   */
-  constructor(dataType: IecType, dimensions: number | number[]) {
-    super()
-
-    this.dataType = dataType
-
-    if (Array.isArray(dimensions)) {
-      this.dimensions = dimensions
-    } else {
-      this.dimensions.push(dimensions)
-    }
-      
-    //Calculating total size
-    this.totalSize = this.dimensions.reduce((total, size) => total * size, 1)
-    this.byteLength = this.totalSize * dataType.byteLength
-  }
-
-
-  convertToBuffer(data: unknown[]): Buffer {
-    if (!data)
-      return Buffer.alloc(0)
-
-    const buffer = Buffer.alloc(this.byteLength)
-    let pos = 0
-
-    for (let i = 0; i < data.length; i++) {
-      const converted = this.dataType.convertToBuffer(data[i])
-
-      converted.copy(buffer, pos)
-      pos += converted.byteLength
-    }
-
-    return buffer
-  }
-
-  convertFromBuffer(data: Buffer): unknown[] {
-    const arr = [] as Array<unknown>
-    let pos = 0
-
-    for (let i = 0; i < this.totalSize; i++) {
-      arr.push(this.dataType.convertFromBuffer(data.slice(pos, pos + this.dataType.byteLength)))
-      pos += this.dataType.byteLength
-    }
-
-    return arr 
-  }
-
-  getDefault(): unknown[] {
-    const arr = [] as Array<unknown>
-
-    for (let i = 0; i < this.totalSize; i++) {
-      arr.push(this.dataType.getDefault())
-    }
-
-    return arr
-  }
-}
-
-
-
+export const ARRAY = (dataType: IecType, dimensions: number | number[]): handler.ARRAY => new handler.ARRAY(dataType, dimensions)
 
 /**
  * IEC 61131-3 type: STRING
  * Default length 80 characters
+ * @param length Length of the string variable (similar as in the PLC), default is 80
  */
-export class STRING extends TypeBase implements IecType {
-  type = 'STRING'
-  byteLength = 80
-
-  /**
-   * Constructor for string
-   * @param length Length of the string variable (similar as in the PLC), default is 80
-   */
-  constructor(length = 80) {
-    super()
-
-    //Adding string end delimeter
-    this.byteLength = length + 1
-  }
-
-  convertToBuffer(data: string): Buffer {
-    const buffer = Buffer.alloc(this.byteLength)
-    iconv.encode(data, 'cp1252').copy(buffer)
-  
-    return buffer
-  }
-
-  convertFromBuffer(data: Buffer): string {
-    return trimPlcString(iconv.decode(data, 'cp1252'))
-  }
-
-  getDefault(): string {
-    return ''
-  }
-}
-
-
-
+export const STRING = (length?: number): handler.STRING => new handler.STRING(length)
 
 /**
  * IEC 61131-3 type: WSTRING
  * Default length 80 characters
+ * @param length Length of the string variable (similar as in the PLC), default is 80
  */
-export class WSTRING extends TypeBase implements IecType {
-  type = 'WSTRING'
-  byteLength = 160
-
-  /**
-   * Constructor for wstring
-   * @param length Length of the string variable (similar as in the PLC), default is 80
-   */
-  constructor(length = 80) {
-    super()
-
-    //Adding string end delimeter
-    this.byteLength = length * 2 + 2
-  }
-
-  convertToBuffer(data: string): Buffer {
-    const buffer = Buffer.alloc(this.byteLength)
-    iconv.encode(data, 'ucs2').copy(buffer)
-
-    return buffer
-  }
-
-  convertFromBuffer(data: Buffer): string {
-    return trimPlcString(iconv.decode(data, 'ucs2'))
-  }
-
-  getDefault(): string {
-    return ''
-  }
-}
-
-
+export const WSTRING = (length?: number): handler.WSTRING => new handler.WSTRING(length)
 
 /**
  * IEC 61131-3 type: BOOL (1 byte)
  */
-export class BOOL extends TypeBase implements IecType {
-  type = 'BOOL'
-  byteLength = 1
-
-  convertToBuffer(data: boolean): Buffer {
-    const buffer = Buffer.alloc(this.byteLength)
-    buffer.writeUInt8(data ? 1 : 0)
-
-    return buffer
-  }
-
-  convertFromBuffer(data: Buffer): boolean {
-    return data.readUInt8() ? true : false
-  }
-
-  getDefault(): boolean {
-    return false
-  }
-}
-
-
+export const BOOL = new handler.BOOL()
 
 /**
  * IEC 61131-3 type: USINT (1 byte)
  */
-export class USINT extends TypeBase implements IecType {
-  type = 'USINT'
-  byteLength = 1
-
-  convertToBuffer(data: number): Buffer {
-    const buffer = Buffer.alloc(this.byteLength)
-    buffer.writeUInt8(data)
-
-    return buffer
-  }
-
-  convertFromBuffer(data: Buffer): number {
-    return data.readUInt8()
-  }
-
-  getDefault(): number {
-    return 0
-  }
-}
-
+export const USINT = new handler.USINT()
 
 /**
  * IEC 61131-3 type: BYTE (1 byte)
  */
-export class BYTE extends USINT implements IecType {
-  type = 'BYTE'
-}
-
-
+export const BYTE = new handler.BYTE()
 
 /**
  * IEC 61131-3 type: SINT (1 byte)
  */
-export class SINT extends TypeBase implements IecType {
-  type = 'SINT'
-  byteLength = 1
-
-  convertToBuffer(data: number): Buffer {
-    const buffer = Buffer.alloc(this.byteLength)
-    buffer.writeInt8(data)
-
-    return buffer
-  }
-
-  convertFromBuffer(data: Buffer): number {
-    return data.readInt8()
-  }
-
-  getDefault(): number {
-    return 0
-  }
-}
-
-
+export const SINT = new handler.SINT()
 
 
 /**
  * IEC 61131-3 type: UINT (2 bytes)
  */
-export class UINT extends TypeBase implements IecType {
-  type = 'UINT'
-  byteLength = 2
-
-  convertToBuffer(data: number): Buffer {
-    const buffer = Buffer.alloc(this.byteLength)
-    buffer.writeUInt16LE(data)
-
-    return buffer
-  }
-
-  convertFromBuffer(data: Buffer): number {
-    return data.readUInt16LE()
-  }
-
-  getDefault(): number {
-    return 0
-  }
-}
-
-
+export const UINT = new handler.UINT()
 
 /**
  * IEC 61131-3 type: WORD (2 bytes)
  */
-export class WORD extends UINT implements IecType {
-  type = 'WORD'
-}
-
-
+export const WORD = new handler.WORD()
 
 /**
  * IEC 61131-3 type: INT (2 bytes)
  */
-export class INT extends TypeBase implements IecType {
-  type = 'INT'
-  byteLength = 2
-
-  convertToBuffer(data: number): Buffer {
-    const buffer = Buffer.alloc(this.byteLength)
-    buffer.writeInt16LE(data)
-
-    return buffer
-  }
-
-  convertFromBuffer(data: Buffer): number {
-    return data.readInt16LE()
-  }
-
-  getDefault(): number {
-    return 0
-  }
-}
+export const INT = new handler.INT()
 
 
 /**
  * IEC 61131-3 type: DINT (4 bytes)
  */
-export class DINT extends TypeBase implements IecType {
-  type = 'DINT'
-  byteLength = 4
-
-  convertToBuffer(data: number): Buffer {
-    const buffer = Buffer.alloc(this.byteLength)
-    buffer.writeInt32LE(data)
-
-    return buffer
-  }
-
-  convertFromBuffer(data: Buffer): number {
-    return data.readInt32LE()
-  }
-
-  getDefault(): number {
-    return 0
-  }
-}
-
-
+export const DINT = new handler.DINT()
 
 /**
  * IEC 61131-3 type: UDINT (4 bytes)
  */
-export class UDINT extends TypeBase implements IecType {
-  type = 'UDINT'
-  byteLength = 4
-
-  convertToBuffer(data: number): Buffer {
-    const buffer = Buffer.alloc(this.byteLength)
-    buffer.writeUInt32LE(data)
-
-    return buffer
-  }
-
-  convertFromBuffer(data: Buffer): number {
-    return data.readUInt32LE()
-  }
-
-  getDefault(): number {
-    return 0
-  }
-}
-
-
+export const UDINT = new handler.UDINT()
 
 /**
  * IEC 61131-3 type: DWORD (4 bytes)
  */
-export class DWORD extends UDINT implements IecType {
-  type = 'DWORD'
-}
-
-
+export const DWORD = new handler.DWORD()
 
 /**
  * IEC 61131-3 type: TIME (4 bytes)
  */
-export class TIME extends UDINT implements IecType {
-  type = 'TIME'
-}
-
-
+export const TIME = new handler.TIME()
 
 /**
  * IEC 61131-3 type: TOD (4 bytes)
  */
-export class TOD extends UDINT implements IecType {
-  type = 'TOD'
-}
-
-
+export const TOD = new handler.TOD()
 
 /**
  * IEC 61131-3 type: TIME_OF_DAY (4 bytes)
  */
-export class TIME_OF_DAY extends TOD implements IecType {
-  type = 'TIME_OF_DAY'
-}
-
-
+export const TIME_OF_DAY = new handler.TIME_OF_DAY()
 
 /**
  * IEC 61131-3 type: DT (4 bytes)
  * TODO: Conversion to Javascript Date object?
  */
-export class DT extends UDINT implements IecType {
-  type = 'DT'
-}
-
-
+export const DT = new handler.DT()
 
 /**
  * IEC 61131-3 type: DATE_AND_TIME (4 bytes)
  * TODO: Conversion to Javascript Date object?
  */
-export class DATE_AND_TIME extends DT implements IecType {
-  type = 'DATE_AND_TIME'
-}
-
-
+export const DATE_AND_TIME = new handler.DATE_AND_TIME()
 
 /**
  * IEC 61131-3 type: DATE (4 bytes)
  * TODO: Conversion to Javascript Date object?
  */
-export class DATE extends UDINT implements IecType {
-  type = 'DATE'
-}
-
-
+export const DATE = new handler.DATE()
 
 
 /**
  * IEC 61131-3 type: REAL (4 bytes)
  */
-export class REAL extends TypeBase implements IecType {
-  type = 'REAL'
-  byteLength = 4
-
-  convertToBuffer(data: number): Buffer {
-    const buffer = Buffer.alloc(this.byteLength)
-    buffer.writeFloatLE(data)
-
-    return buffer
-  }
-
-  convertFromBuffer(data: Buffer): number {
-    return data.readFloatLE()
-  }
-
-  getDefault(): number {
-    return 0
-  }
-}
-
-
-
+export const REAL = new handler.REAL()
 
 /**
  * IEC 61131-3 type: LREAL (4 bytes)
  */
-export class LREAL extends TypeBase implements IecType {
-  type = 'LREAL'
-  byteLength = 8
-
-  convertToBuffer(data: number): Buffer {
-    const buffer = Buffer.alloc(this.byteLength)
-    buffer.writeDoubleLE(data)
-
-    return buffer
-  }
-
-  convertFromBuffer(data: Buffer): number {
-    return data.readDoubleLE()
-  }
-
-  getDefault(): number {
-    return 0
-  }
-}
-
-
+export const LREAL = new handler.LREAL()
 
 
 /**
  * IEC 61131-3 type: ULINT (8 bytes)
  * TODO: Requires Node.js that supports BigInt
  */
-export class ULINT extends TypeBase implements IecType {
-  type = 'ULINT'
-  byteLength = 8
-
-  convertToBuffer(data: bigint): Buffer {
-    const buffer = Buffer.alloc(this.byteLength)
-    buffer.writeBigUInt64LE(data)
-
-    return buffer
-  }
-
-  convertFromBuffer(data: Buffer): bigint {
-    return data.readBigUInt64LE()
-  }
-
-  getDefault(): number {
-    return 0
-  }
-}
-
-
-
+export const ULINT = new handler.ULINT()
 
 /**
  * IEC 61131-3 type: LWORD (8 bytes)
  * TODO: Requires Node.js that supports BigInt
  */
-export class LWORD extends ULINT implements IecType {
-  type = 'LWORD'
-}
-
-
-
+export const LWORD = new handler.LWORD()
 
 /**
  * IEC 61131-3 type: LINT (8 bytes)
  * TODO: Requires Node.js that supports BigInt
  */
-export class LINT extends TypeBase implements IecType {
-  type = 'LINT'
-  byteLength = 8
-
-  convertToBuffer(data: bigint): Buffer {
-    const buffer = Buffer.alloc(this.byteLength)
-    buffer.writeBigInt64LE(data)
-
-    return buffer
-  }
-
-  convertFromBuffer(data: Buffer): bigint {
-    return data.readBigInt64LE()
-  }
-
-  getDefault(): number {
-    return 0
-  }
-}
+export const LINT = new handler.LINT()
 
 
 
 
 
-/**
- * Trims the given PLC string until end mark (\0, 0 byte) is found
- * (= removes empty bytes from end of the string)
- * @param {string} plcString String to trim
- * 
- * @returns {string} Trimmed string
- */
-const trimPlcString = (plcString: string): string => {
-  let parsedStr = ''
 
-  for (let i = 0; i < plcString.length; i++) {
-    if (plcString.charCodeAt(i) === 0) break
-
-    parsedStr += plcString[i]
-  }
-
-  return parsedStr
-}
-
-
+
